tests/wpt/tests/css/motion/animation/ray-angle-interpolation-math-functions.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8">
    <title>ray() angle interpolation with css math functions</title>
    <link rel="author" title="Daniil Sakhapov" href="mailto:sakhapov@chromium.org">
    <link rel="help" href="https://drafts.fxtf.org/motion/#offset-path-property">
    <link rel="help" href="https://drafts.csswg.org/css-values/#sign-funcs">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/css/support/interpolation-testcommon.js"></script>
  </head>
  <style>
    :root {
      font-size: 10px;
    }
  </style>
  <body>
    <script>
      test_interpolation({
        property: 'offset-path',
        from: 'ray(100deg)',
        to: 'ray(calc(sign(20rem - 20px) * 180deg))',
      }, [
        {at: -1, expect: 'ray(20deg)'},
        {at: 0, expect: 'ray(100deg)'},
        {at: 0.125, expect: 'ray(110deg)'},
        {at: 0.875, expect: 'ray(170deg)'},
        {at: 1, expect: 'ray(180deg)'},
        {at: 2, expect: 'ray(260deg)'}
      ]);

      test_interpolation({
        property: 'offset-path',
        from: 'ray(calc(sign(20rem - 20px) * 100deg))',
        to: 'ray(calc(sign(20rem - 20px) * 180deg))',
      }, [
        {at: -1, expect: 'ray(20deg)'},
        {at: 0, expect: 'ray(100deg)'},
        {at: 0.125, expect: 'ray(110deg)'},
        {at: 0.875, expect: 'ray(170deg)'},
        {at: 1, expect: 'ray(180deg)'},
        {at: 2, expect: 'ray(260deg)'}
      ]);
    </script>
  </body>
</html>
