LayoutTests/imported/w3c/web-platform-tests/css/motion/animation/ray-angle-interpolation-math-functions-expected.txt


FAIL CSS Transitions: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (-1) should be [ray(20deg)] assert_equals: expected "ray ( 20deg ) " but got "ray ( 200deg ) "
PASS CSS Transitions: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0) should be [ray(100deg)]
FAIL CSS Transitions: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.125) should be [ray(110deg)] assert_equals: expected "ray ( 110deg ) " but got "ray ( 87.5deg ) "
FAIL CSS Transitions: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.875) should be [ray(170deg)] assert_equals: expected "ray ( 170deg ) " but got "ray ( 12.5deg ) "
FAIL CSS Transitions: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (1) should be [ray(180deg)] assert_equals: expected "ray ( 180deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (2) should be [ray(260deg)] assert_equals: expected "ray ( 260deg ) " but got "ray ( - 100deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (-1) should be [ray(20deg)] assert_equals: expected "ray ( 20deg ) " but got "ray ( 200deg ) "
PASS CSS Transitions with transition: all: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0) should be [ray(100deg)]
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.125) should be [ray(110deg)] assert_equals: expected "ray ( 110deg ) " but got "ray ( 87.5deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.875) should be [ray(170deg)] assert_equals: expected "ray ( 170deg ) " but got "ray ( 12.5deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (1) should be [ray(180deg)] assert_equals: expected "ray ( 180deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (2) should be [ray(260deg)] assert_equals: expected "ray ( 260deg ) " but got "ray ( - 100deg ) "
FAIL CSS Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (-1) should be [ray(20deg)] assert_equals: expected "ray ( 20deg ) " but got "ray ( 200deg ) "
PASS CSS Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0) should be [ray(100deg)]
FAIL CSS Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.125) should be [ray(110deg)] assert_equals: expected "ray ( 110deg ) " but got "ray ( 87.5deg ) "
FAIL CSS Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.875) should be [ray(170deg)] assert_equals: expected "ray ( 170deg ) " but got "ray ( 12.5deg ) "
FAIL CSS Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (1) should be [ray(180deg)] assert_equals: expected "ray ( 180deg ) " but got "ray ( 0deg ) "
FAIL CSS Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (2) should be [ray(260deg)] assert_equals: expected "ray ( 260deg ) " but got "ray ( - 100deg ) "
FAIL Web Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (-1) should be [ray(20deg)] assert_equals: expected "ray ( 20deg ) " but got "ray ( 200deg ) "
PASS Web Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0) should be [ray(100deg)]
FAIL Web Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.125) should be [ray(110deg)] assert_equals: expected "ray ( 110deg ) " but got "ray ( 87.5deg ) "
FAIL Web Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.875) should be [ray(170deg)] assert_equals: expected "ray ( 170deg ) " but got "ray ( 12.5deg ) "
FAIL Web Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (1) should be [ray(180deg)] assert_equals: expected "ray ( 180deg ) " but got "ray ( 0deg ) "
FAIL Web Animations: property <offset-path> from [ray(100deg)] to [ray(calc(sign(20rem - 20px) * 180deg))] at (2) should be [ray(260deg)] assert_equals: expected "ray ( 260deg ) " but got "ray ( - 100deg ) "
FAIL CSS Transitions: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (-1) should be [ray(20deg)] assert_equals: expected "ray ( 20deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0) should be [ray(100deg)] assert_equals: expected "ray ( 100deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.125) should be [ray(110deg)] assert_equals: expected "ray ( 110deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.875) should be [ray(170deg)] assert_equals: expected "ray ( 170deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (1) should be [ray(180deg)] assert_equals: expected "ray ( 180deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (2) should be [ray(260deg)] assert_equals: expected "ray ( 260deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (-1) should be [ray(20deg)] assert_equals: expected "ray ( 20deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0) should be [ray(100deg)] assert_equals: expected "ray ( 100deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.125) should be [ray(110deg)] assert_equals: expected "ray ( 110deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.875) should be [ray(170deg)] assert_equals: expected "ray ( 170deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (1) should be [ray(180deg)] assert_equals: expected "ray ( 180deg ) " but got "ray ( 0deg ) "
FAIL CSS Transitions with transition: all: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (2) should be [ray(260deg)] assert_equals: expected "ray ( 260deg ) " but got "ray ( 0deg ) "
FAIL CSS Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (-1) should be [ray(20deg)] assert_equals: expected "ray ( 20deg ) " but got "ray ( 0deg ) "
FAIL CSS Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0) should be [ray(100deg)] assert_equals: expected "ray ( 100deg ) " but got "ray ( 0deg ) "
FAIL CSS Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.125) should be [ray(110deg)] assert_equals: expected "ray ( 110deg ) " but got "ray ( 0deg ) "
FAIL CSS Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.875) should be [ray(170deg)] assert_equals: expected "ray ( 170deg ) " but got "ray ( 0deg ) "
FAIL CSS Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (1) should be [ray(180deg)] assert_equals: expected "ray ( 180deg ) " but got "ray ( 0deg ) "
FAIL CSS Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (2) should be [ray(260deg)] assert_equals: expected "ray ( 260deg ) " but got "ray ( 0deg ) "
FAIL Web Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (-1) should be [ray(20deg)] assert_equals: expected "ray ( 20deg ) " but got "ray ( 0deg ) "
FAIL Web Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0) should be [ray(100deg)] assert_equals: expected "ray ( 100deg ) " but got "ray ( 0deg ) "
FAIL Web Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.125) should be [ray(110deg)] assert_equals: expected "ray ( 110deg ) " but got "ray ( 0deg ) "
FAIL Web Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (0.875) should be [ray(170deg)] assert_equals: expected "ray ( 170deg ) " but got "ray ( 0deg ) "
FAIL Web Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (1) should be [ray(180deg)] assert_equals: expected "ray ( 180deg ) " but got "ray ( 0deg ) "
FAIL Web Animations: property <offset-path> from [ray(calc(sign(20rem - 20px) * 100deg))] to [ray(calc(sign(20rem - 20px) * 180deg))] at (2) should be [ray(260deg)] assert_equals: expected "ray ( 260deg ) " but got "ray ( 0deg ) "
